add new files

+++ GA.TradeMarket.Domain/Interfaces/ICrudRep.cs
@@ -8,15 +8,15 @@ namespace GA.TradeMarket.Domain.Interfaces
 {
     public interface ICrudRep<T> where T : class
     {
-        Task<T> GetByIdAsync(int Id);
+        Task<T> GetByIdAsync(long Id);
 
         Task<IEnumerable<T>> GetAllAsync();
 
         Task AddAsync(T customer);
 
-        Task DeleteByIdAsync(int Id);
+        Task DeleteByIdAsync(long Id);
 
-        Task<T> GetByIdWithDetailsAsync(int Id);
+        Task<T> GetByIdWithDetailsAsync(long Id);
 
         Task<IEnumerable<T>> GetAllWithDetailsAsync();
