+++ GA.TradeMarket.Infrastructure/Repositories/ProductCategoryRepository .cs
@@ -28,7 +28,7 @@ namespace GA.TradeMarket.Infrastructure.Repositories
             }
         }
 
-        public async Task DeleteByIdAsync(int Id)
+        public async Task DeleteByIdAsync(long Id)
         {
             var res = await dbset.FirstOrDefaultAsync(io => io.Id == Id);
             if (res is not null)
@@ -49,7 +49,7 @@ namespace GA.TradeMarket.Infrastructure.Repositories
             return await dbset.Include(io => io.Products).ToListAsync();
         }
 
-        public async Task<ProductCategory> GetByIdAsync(int Id)
+        public async Task<ProductCategory> GetByIdAsync(long Id)
         {
             var res = await dbset.FirstOrDefaultAsync(io => io.Id == Id);
             if (res is not  null)
@@ -59,7 +59,7 @@ namespace GA.TradeMarket.Infrastructure.Repositories
             throw new ArgumentException("No entity DOund on this ID");
         }
 
-        public async Task<ProductCategory> GetByIdWithDetailsAsync(int Id)
+        public async Task<ProductCategory> GetByIdWithDetailsAsync(long Id)
         {
             var res = await dbset.Include(io => io.Products).FirstOrDefaultAsync(io => io.Id == Id);
             if (res != null)
